backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
alembic==1.16.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic-settings==2.10.1
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
resend==2.13.0
pytest==8.4.1
pytest-asyncio==1.1.0
